requirements.txt
scrapy
kafka-python